+++ src/main/java/com/qoormthon/empty_wallet/domain/user/entity/Gender.java
@@ -1,4 +1,4 @@
-package com.qoormthon.empty_wallet.domain.user.dto;
+package com.qoormthon.empty_wallet.domain.user.entity;
 
 public enum Gender {
